net/i40e: check register read for outer VLAN
[ upstream commit c11c52dd5d2a19c97616ac32a1d4911c48f157d4 ]

'i40e_get_outer_vlan()' does not check 'i40e_aq_debug_read_register()'
return value. This patch fixes this issue, by checking the return value
and, on error, having the i40e_get_outer_vlan() function return that
error back to the caller.

This in turn requires a change in the return type of that function and
updates to the places where it is called to:
* handle the error, and
* handle the tpid being returned as an "out" parameter rather than
  return code.

Coverity issue: 445518
Fixes: 86eb05d6350b ("net/i40e: add flow validate function")

Signed-off-by: Vladimir Medvedkin <vladimir.medvedkin@intel.com>
Acked-by: Bruce Richardson <bruce.richardson@intel.com>